Ajout de l'envoi des messages via SigFox. Ajout de code de test pour l'envoi des données vers le serveur NodeRed via SigFox.

+++ include/sensors/LuxSensor.h
@@ -34,7 +34,7 @@ public:
                                 sensor(Adafruit_VEML7700())
     {
     }
-    long getLuxMeasureAndStop();
+    uint8_t getLuxMeasureAndStop();
 };
 
 #endif
